BC-breaking reduction of package responsibilities.
* * *

Logic for loading project and package configs is now *completely* different.  Previously, we used `define.php` and `modify.php` include files in a directory named for the config mode, and used an Includer to load them via glob().

    vendor/
        packages/
            config/
                dev/
                    define.php
                    modify.php
                prod/
                    define.php
                    modify.php

Now, we expect `Aura\Di\Config` classes mapped to the config mode. The config classes use the package namespace with a sub-namespace of `_Config`; the class name is the config mode.

    vendor/
        package/
            config/
                Common.php # Vendor\\Package\\_Config\\Common
                Dev.php    # Vendor\\Package\\_Config\\Dev
                Prod.php   # Vendor\\Package\\_Config\\Prod

The config classes are explicitly noted in the composer.json file under a new `{"aura": {"extra": {"config": {...}}}}` element, and are mapped for autoload via a PSR-4 entry in `composer.json`.

    "autoload": {
        "psr-4": {
            "Vendor\\Package\\": "src/",
            "Vendor\\Package\\_Config\\": "config/"
        }
    }
    "extra": {
        "aura": {
            "type": "library",
            "config": {
                "common": "Vendor\\Package\\_Config\\Common",
                "dev": "Vendor\\Package\\_Config\\Dev",
                "prod": "Vendor\\Package\\_Config\\Prod",
            }
        }
    }

This change has the benefit of avoiding glob() and therefore file system scans, using the autoloader instead of includes. It also leverages Composer to be more explicit about config information, rather than implicitly scanning for files.

Instead of "default" being the config that always gets loaded, it is now "common" that always gets loaded. This is because "default" is a PHP keyword and cannot be used as a class name.

* * *

Change list:

- No longer requires Monolog/Monolog; we need to add this at the project level.

- No longer sets a 'logger' service; we need to add this at the project level.

- No longer has any `config/` elements.

- No longer has any `script/` files -- kernel setup will short enough (about 4 calls) to copy over at the project level.

- Config information is now defined as class names in the Composer `extra->aura->config` object. The field name is the config mode, and the value is the class to use for configuration. Use Project::getConfigClasses() to retrieve the list of classes in their proper order (i.e., libraries, kernels, projects).

- The Project object has no more directory-specific methods, e.g. getTmpPath() is now gone. Use getPath() instead.

- The Project object now retains both `installed.json` *and* the project-level `composer.json` data.

- Container define/modify cycle is moved to Aura\Di\ContainerBuilder.

- ProjectContainer renamed to Factory, and creates the project container via Aura\Di\ContainerBuilder.

- Discovery of config classes is now part of the Project object.

- Integration testing no longer requires a *_Project; the tests now need only an Aura\Di installation via Composer. Continue using `tests/integration.sh` to run the tests.

- Now runs under PHP 5.3.